LayoutTests/storage/indexeddb/versionchangerequest-activedomobject-expected.txt

CONSOLE MESSAGE: The setVersion() method is non-standard and will be removed. Use the "upgradeneeded" event instead.
Ensure that IDBVersionChangeRequest objects are not GC'd if they have pending events

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;


self.dbname = 'versionchangerequest-activedomobject'

testDeleteDatabase():
indexedDB.deleteDatabase(self.dbname)
self.gc()
PASS deleteDatabase's IDBVersionChangeRequest.onsuccess event fired

testSetVersion():
openRequest = indexedDB.open(self.dbname)
db = openRequest.result
db.setVersion('1')
self.gc()
PASS setVersion's IDBVersionChangeRequest.onsuccess event fired
PASS successfullyParsed is true

TEST COMPLETE
